refactor: extract logic to config manager
